refactor: 디렉토리명을 logger에서 loggers로 변경
