+++ packages/logger/src/index.ts
@@ -1,4 +1,4 @@
-export { default as logger } from './logger';
+export { default as logger } from './loggers';
 export { FirebaseConfig } from './models/firebase';
 export {
   LoggerEnvMode,
