+++ packages/logger/src/loggers/firebase/index.ts
